# Improve return configuration