--- a/src/NSubstitute/Routing/Handlers/ReturnFromAndConfigureDynamicCall.cs
+++ b/src/NSubstitute/Routing/Handlers/ReturnFromAndConfigureDynamicCall.cs
@@ -52,22 +52,22 @@ namespace NSubstitute.Routing.Handlers
         {
             public ConfiguredCall Returns<T>(T returnThis, params T[] returnThese)
             {
-                return SubstituteExtensions.Returns(MatchArgs.AsSpecifiedInCall, returnThis, returnThese);
+                return default(T).Returns(returnThis, returnThese);
             }
 
             public ConfiguredCall Returns<T>(Func<CallInfo, T> returnThis, params Func<CallInfo, T>[] returnThese)
             {
-                return SubstituteExtensions.Returns(MatchArgs.AsSpecifiedInCall, returnThis, returnThese);
+                return default(T).Returns(returnThis, returnThese);
             }
 
             public ConfiguredCall ReturnsForAnyArgs<T>(T returnThis, params T[] returnThese)
             {
-                return SubstituteExtensions.Returns(MatchArgs.Any, returnThis, returnThese);
+                return default(T).ReturnsForAnyArgs(returnThis, returnThese);
             }
 
             public ConfiguredCall ReturnsForAnyArgs<T>(Func<CallInfo, T> returnThis, params Func<CallInfo, T>[] returnThese)
             {
-                return SubstituteExtensions.Returns(MatchArgs.Any, returnThis, returnThese);
+                return default(T).ReturnsForAnyArgs(returnThis, returnThese);
             }
         }
     }