Tiny performance and allocation improvement for the case when we want to return the already configured value. I found that  we don't need information about the arguments when we want return a static value. Again, it's a small improvement, but now we need less than 1 ns - isn't it wonderful (and, of course, fully useless in real life)? :sweat_smile:

Bechmark (tested only one case, as other would produce the same value):

```
BEFORE

                         Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
------------------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
 DispatchInterfaceProxyCall_Int | 1.044 us | 0.0205 us | 0.0300 us | 0.0801 | 0.0248 |     536 B |
 

AFTER:

 DispatchInterfaceProxyCall_Int | 958.0 ns | 15.60 ns | 13.03 ns | 0.0620 | 0.0181 |     408 B |

```

Additional, I slightly refactored out `Return` configuration extensions, so now they are more consistent.

@dtchepak Happy review, let me know if you have any objections.